+++ custom_components/homehoard/__init__.py
@@ -5,14 +5,14 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers import aiohttp_client
 
 from .const import DOMAIN
-from .coordinator import ShelfieDataUpdateCoordinator
+from .coordinator import HomeHoardDataUpdateCoordinator
 
 PLATFORMS = ["sensor"]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     session = aiohttp_client.async_get_clientsession(hass)
-    coordinator = ShelfieDataUpdateCoordinator(hass, session, entry)
+    coordinator = HomeHoardDataUpdateCoordinator(hass, session, entry)
     await coordinator.async_config_entry_first_refresh()
 
     hass.data.setdefault(DOMAIN, {})[entry.entry_id] = coordinator